test(integration): move controller batch 4
